+++ eras/conway/impl/CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ## 1.23.0.0
 
+* Add `ToJSON` instance for `DefaultVote`.
 * Add `injectStakeCredentials`, `injectDRepsThenDelegs` to `Cardano.Ledger.Conway.Transition`
 * Add `ConwayExtraConfig` type and `cgExtraConfig` field to `ConwayGenesis`
 * Deprecate `conwayRegisterInitialAccounts`, `registerDRepsThenDelegs`